Added the support for push notifications

+++ db/database.go
@@ -10,6 +10,7 @@ import (
 	"github.com/desmos-labs/juno/config"
 	"github.com/desmos-labs/juno/db"
 	"github.com/desmos-labs/juno/db/postgresql"
+	"github.com/jmoiron/sqlx"
 	"github.com/lib/pq"
 )
 
@@ -17,6 +18,7 @@ import (
 // so that it can properly store posts and other Desmos-related data.
 type DesmosDb struct {
 	postgresql.Database
+	sqlx *sqlx.DB
 }
 
 // Builder allows to create a new DesmosDb instance implementing the database.Builder type
@@ -27,7 +29,11 @@ func Builder(cfg config.Config, codec *codec.Codec) (*db.Database, error) {
 	}
 
 	psqlDb, _ := (*database).(postgresql.Database)
-	var desmosDb db.Database = DesmosDb{Database: psqlDb}
+	var desmosDb db.Database = DesmosDb{
+		Database: psqlDb,
+		sqlx:     sqlx.NewDb(psqlDb.Sql, "postgresql"),
+	}
+
 	return &desmosDb, nil
 }
 
@@ -35,6 +41,7 @@ func Builder(cfg config.Config, codec *codec.Codec) (*db.Database, error) {
 func (db DesmosDb) SavePost(post posts.Post) error {
 	var pollID *uint64
 
+	// TODO: Split this
 	if post.PollData != nil {
 		// Saving post's poll data before post to make possible the insertion of poll_id inside it
 		statement := `
@@ -53,11 +60,7 @@ func (db DesmosDb) SavePost(post posts.Post) error {
 			return err
 		}
 
-		pollQuery := `
-		INSERT INTO poll_answer(poll_id, answer_id, answer_text)
-		VALUES($1, $2, $3)
-		`
-
+		pollQuery := `INSERT INTO poll_answer(poll_id, answer_id, answer_text) VALUES($1, $2, $3)`
 		for _, answer := range post.PollData.ProvidedAnswers {
 			_, err := db.Sql.Exec(pollQuery, pollID, answer.ID, answer.Text)
 			if err != nil {
@@ -73,10 +76,12 @@ func (db DesmosDb) SavePost(post posts.Post) error {
 	VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10)
     `
 
-	// todo look how this is inserted in DB
-	jsonB, _ := json.Marshal(post.OptionalData)
+	jsonB, err := json.Marshal(post.OptionalData)
+	if err != nil {
+		return err
+	}
 
-	_, err := db.Sql.Exec(
+	_, err = db.Sql.Exec(
 		postSqlStatement,
 		post.PostID, post.ParentID, post.Message, post.Created, post.LastEdited, post.AllowsComments, post.Subspace,
 		post.Creator.String(), pollID, string(jsonB),
@@ -86,11 +91,8 @@ func (db DesmosDb) SavePost(post posts.Post) error {
 	}
 
 	// Saving post's medias
-	mediaQuery := `
-	INSERT INTO media (post_id, uri, mime_type)
-	VALUES ($1, $2, $3)
-	`
-
+	// TODO: Split this
+	mediaQuery := `INSERT INTO media (post_id, uri, mime_type) VALUES ($1, $2, $3)`
 	for _, media := range post.Medias {
 		_, err = db.Sql.Exec(mediaQuery, post.PostID, media.URI, media.MimeType)
 		if err != nil {
@@ -104,35 +106,42 @@ func (db DesmosDb) SavePost(post posts.Post) error {
 // EditPost allows to properly edit the post having the given postID by setting the new
 // given message and editDate
 func (db DesmosDb) EditPost(postID posts.PostID, message string, editDate time.Time) error {
-	statement := `
-	UPDATE post 
-	SET message = $1, last_edited = $2 
-	WHERE id = $3
-	`
-
+	statement := `UPDATE post SET message = $1, last_edited = $2 WHERE id = $3`
 	_, err := db.Sql.Exec(statement, message, editDate, postID)
 	return err
 }
 
-// SaveReaction allows to save a new reaction for the given postID having the specified value and user
-func (db DesmosDb) SaveReaction(postID posts.PostID, reaction posts.Reaction) error {
-	statement := `
-	INSERT INTO reaction (post_id, owner, value)
-	VALUES ($1, $2, $3)
-	`
+// GetPostByID returns the post having the specified id.
+// If some error raised during the read, it is returned.
+// If no post with the specified id is found, nil is returned instead.
+func (db DesmosDb) GetPostByID(id posts.PostID) (*posts.Post, error) {
+	postSqlStatement := `SELECT * FROM post WHERE id = $1`
 
+	var rows []PostRow
+	err := db.sqlx.Select(&rows, postSqlStatement, id)
+	if err != nil {
+		return nil, err
+	}
+
+	// No post found
+	if len(rows) == 0 {
+		return nil, nil
+	}
+
+	return ConvertPostRow(rows[0])
+}
+
+// SaveReaction allows to save a new reaction for the given postID having the specified value and user
+func (db DesmosDb) SaveReaction(postID posts.PostID, reaction posts.Reaction) (*posts.Reaction, error) {
+	statement := `INSERT INTO reaction (post_id, owner, value) VALUES ($1, $2, $3)`
 	_, err := db.Sql.Exec(statement, postID, reaction.Owner.String(), reaction.Value)
-	return err
+	return &reaction, err
 }
 
 // RemoveReaction allows to remove an already existing reaction for the post having the given postID,
 // the given reaction and from the specified user.
 func (db DesmosDb) RemoveReaction(postID posts.PostID, reaction string, user sdk.AccAddress) error {
-	statement := `
-	DELETE FROM reaction
-	WHERE post_id = $1 AND owner = $2 AND reaction = $3;
-	`
-
+	statement := `DELETE FROM reaction WHERE post_id = $1 AND owner = $2 AND reaction = $3;`
 	_, err := db.Sql.Exec(statement, postID, user.String(), reaction)
 	return err
 }
@@ -140,11 +149,7 @@ func (db DesmosDb) RemoveReaction(postID posts.PostID, reaction string, user sdk
 // SavePollAnswer allows to save the given answers from the specified user for the poll
 // post having the specified postID.
 func (db DesmosDb) SavePollAnswer(postID posts.PostID, answer posts.UserAnswer) error {
-	statement := `
-	INSERT INTO user_poll_answer (poll_id, answers, user_address)
-	VALUES ($1, $2, $3)
-	`
-
+	statement := `INSERT INTO user_poll_answer (poll_id, answers, user_address) VALUES ($1, $2, $3)`
 	_, err := db.Sql.Exec(statement, postID, pq.Array(answer.Answers), answer.User.String())
 	return err
 }
